+++ src/chart/ui/events.cpp
@@ -6,7 +6,7 @@
 using namespace Vizzu;
 using namespace Vizzu::UI;
 
-ClickEvent::ClickEvent(Geom::Point position,
+MouseEvent::MouseEvent(Geom::Point position,
 	const Diag::Marker *marker,
 	Chart &chart) :
     Util::EventDispatcher::Params(&chart),
@@ -15,36 +15,21 @@ ClickEvent::ClickEvent(Geom::Point position,
 {
 }
 
-std::string ClickEvent::dataToJson() const
+std::string MouseEvent::dataToJson() const
 {
 	std::string markerJson;
-	if (marker) {
-		const auto &chart = *dynamic_cast<const Vizzu::Chart*>(sender);
-		auto diagram = chart.getDiagram();
-		if (diagram)
-			markerJson = marker->toJson(diagram->getDataCube());
+	auto coords = Geom::Point::Invalid();
+	const auto &chart = *dynamic_cast<const Vizzu::Chart*>(sender);
+	auto diagram = chart.getDiagram();
+	if (diagram) {
+		if (marker) markerJson = marker->toJson(diagram->getDataCube());
+		coords = chart.getCoordSystem().getOriginal(position);
 	}
 	return
 		"{"
-			"\"position\":" + std::string(position) + ", "
+			"\"position\":" + std::string(position)
+			+ ",\"coords\":" + std::string(coords)
+			+ (!markerJson.empty() ? ", ": "")
 			+ markerJson +
 		"}";
-}
-
-MouseOnEvent::MouseOnEvent(Chart &chart, const Diag::Marker *marker) :
-    Util::EventDispatcher::Params(&chart), marker(marker)
-{
-}
-
-std::string MouseOnEvent::dataToJson() const
-{
-	std::string markerJson;
-	if (marker) {
-		const auto &chart = *dynamic_cast<const Vizzu::Chart*>(sender);
-		auto diagram = chart.getDiagram();
-		if (diagram)
-			markerJson = marker->toJson(diagram->getDataCube());
-		return "{" + markerJson + "}";
-	}
-	return "\"\"";
 }